Add public Group32 interface and an implementation for dist1234.
group_impl module contains generic implementations of these methods that
can be used with any UnsafeGroup implementation, which should minimize
errors and code duplication. The encode and decode methods within
group_impl are templatized to allow injecting transformation code to
maximize sharing there as well.

Group64 has been deliberately omitted. This could take the same form as
it does in UnsafeGroup with the Elem num_traits type, but I don't want
to expose this directly in the interface.