Set configuration to target ssse3 by default on x86_64 targets (#5)
Target this feature as we'd like to generate native instructions
(instead of emulated calls) in all cases. There is a runtime
switch to prevent use of ssse3 code paths if those instructions are not
available. This also ensures that we will run
appropriate tests on hosts that have the feature.